Separate out `tallying.rs`
